fix[ping]: start ping if not received msg in time
There are cases that the upper layer router or isp prevents a message
from being sent. If it happend, we can't send message to server, but
from the mqtt process's point of view, no error has occurred. So, we
have to start ping detection if we didn't receive mqtt message within
the ping interval.

It happened to some user, in this case, the home center will not be able
to detect offline and takes longer to reconnect.